+++ cmsplugin_cascade/bootstrap3/gallery.py
@@ -177,7 +177,7 @@ class BootstrapGalleryPlugin(CascadePluginBase):
 
     @classmethod
     def get_css_classes(cls, obj):
-        css_classes = super(BootstrapGalleryPlugin, cls).get_css_classes(obj)
+        css_classes = cls.super(BootstrapGalleryPlugin, cls).get_css_classes(obj)
         css_class = obj.glossary.get('css_class')
         if css_class:
             css_classes.append(css_class)