+++ cmsplugin_cascade/icon/cms_plugins.py
@@ -18,7 +18,7 @@ from .mixins import IconPluginMixin, IconModelMixin
 
 
 class FramedIconPlugin(IconPluginMixin, CascadePluginBase):
-    name = _("Icon")
+    name = _("Icon with frame")
     parent_classes = None
     require_parent = False
     allow_children = False
@@ -88,7 +88,7 @@ class FramedIconPlugin(IconPluginMixin, CascadePluginBase):
 
     @classmethod
     def get_css_classes(cls, instance):
-        css_classes = super(FramedIconPlugin, cls).get_css_classes(instance)
+        css_classes = cls.super(FramedIconPlugin, cls).get_css_classes(instance)
         text_align = instance.glossary.get('text_align')
         if text_align:
             css_classes.append(text_align)
@@ -96,7 +96,7 @@ class FramedIconPlugin(IconPluginMixin, CascadePluginBase):
 
     @classmethod
     def get_inline_styles(cls, instance):
-        inline_styles = super(FramedIconPlugin, cls).get_inline_styles(instance)
+        inline_styles = cls.super(cls, cls).get_inline_styles(instance)
         inline_styles['font-size'] = instance.glossary.get('font_size', '1em')
         return inline_styles
 
@@ -114,7 +114,7 @@ class TextIconModelMixin(object):
 
 
 class TextIconPlugin(IconPluginMixin, CascadePluginBase):
-    name = _("Icon")
+    name = _("Icon in text")
     text_enabled = True
     render_template = 'cascade/plugins/texticon.html'
     ring_plugin = 'IconPlugin'