+++ cmsplugin_cascade/plugin_base.py
@@ -2,7 +2,6 @@
 from __future__ import unicode_literals
 
 from collections import OrderedDict
-from distutils.version import LooseVersion
 
 from django.core.exceptions import ImproperlyConfigured, ObjectDoesNotExist
 from django.forms import MediaDefiningClass
@@ -12,16 +11,17 @@ from django.utils.module_loading import import_string
 from django.utils.translation import string_concat
 from django.utils.safestring import SafeText, mark_safe
 
-from cms import __version__ as cms_version
 from cms.plugin_base import CMSPluginBaseMetaclass, CMSPluginBase
 from cms.utils.compat.dj import is_installed
 
 from . import app_settings
 from .fields import GlossaryField
+from .mixins import CascadePluginMixin
 from .models_base import CascadeModelBase
 from .models import CascadeElement, SharableCascadeElement
 from .generic.mixins import SectionMixin, SectionModelMixin
 from .sharable.forms import SharableGlossaryMixin
+from .strides import register_stride
 from .extra_fields.mixins import ExtraFieldsMixin
 from .widgets import JSONMultiWidget
 from .hide_plugins import HidePluginMixin
@@ -150,8 +150,13 @@ class CascadePluginBaseMetaclass(CascadePluginMixinMetaclass, CMSPluginBaseMetac
         if name == 'SegmentPlugin':
             # SegmentPlugin shall additionally inherit from configured mixin classes
             model_mixins += tuple(import_string(mc[0]) for mc in app_settings.CMSPLUGIN_CASCADE['segmentation_mixins'])
-        module = attrs.get('__module__')
-        attrs['model'] = create_proxy_model(name, model_mixins, base_model, module=module)
+        if 'model' in attrs:
+            # the plugin overrides the CascadeModel
+            if not issubclass(attrs['model'], CascadeModelBase):
+                msg = "Cascade Plugins, overriding the model, must inherit from `CascadeModelBase`."
+                raise ImproperlyConfigured(msg)
+        else:
+            attrs['model'] = create_proxy_model(name, model_mixins, base_model, module=attrs.get('__module__'))
         if is_installed('reversion'):
             import reversion.revisions
             if not reversion.revisions.is_registered(base_model):
@@ -161,6 +166,9 @@ class CascadePluginBaseMetaclass(CascadePluginMixinMetaclass, CMSPluginBaseMetac
             attrs['name'] = mark_safe_lazy(string_concat(
                 app_settings.CMSPLUGIN_CASCADE['plugin_prefix'], "&nbsp;", attrs['name']))
 
+        register_stride(name, bases, attrs, model_mixins)
+        if name == 'CascadePluginBase':
+            bases += (CascadePluginMixin, CMSPluginBase,)
         return super(CascadePluginBaseMetaclass, cls).__new__(cls, name, bases, attrs)
 
 
@@ -228,7 +236,7 @@ class TransparentContainer(TransparentWrapper):
             return _leaf_transparent_plugins
 
 
-class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPluginBase)):
+class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass)):
     change_form_template = 'cascade/admin/change_form.html'
     glossary_variables = []  # entries in glossary not handled by a form editor
     model_mixins = ()  # model mixins added to the final Django model
@@ -248,6 +256,16 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
 
     def __repr__(self):
         return "<class '{}'>".format(self.__class__.__name__)
+
+    @classmethod
+    def super(cls, klass, instance):
+        """
+        Plugins inheriting from CascadePluginBaseMetaclass can have two different base classes,
+        :class:`cmsplugin_cascade.plugin_base.CMSPluginBase` and :class:`cmsplugin_cascade.strides.StridePluginBase`.
+        Therefore in order to call a method from a inherited class, use this "super" wrapping method.
+        >>> cls.super(MyPlugin, self).a_method()
+        """
+        return super(klass, instance)
 
     @classmethod
     def _get_parent_classes_transparent(cls, slot, page, instance=None):
@@ -297,51 +315,6 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
         return SafeText()
 
     @classmethod
-    def get_tag_type(self, instance):
-        """
-        Return the tag_type used to render this plugin.
-        """
-        return instance.glossary.get('tag_type', getattr(self, 'tag_type', 'div'))
-
-    @classmethod
-    def get_css_classes(cls, instance):
-        """
-        Returns a list of CSS classes to be added as class="..." to the current HTML tag.
-        """
-        css_classes = []
-        if hasattr(cls, 'default_css_class'):
-            css_classes.append(cls.default_css_class)
-        for attr in getattr(cls, 'default_css_attributes', []):
-            css_class = instance.glossary.get(attr)
-            if isinstance(css_class, six.string_types):
-                css_classes.append(css_class)
-            elif isinstance(css_class, list):
-                css_classes.extend(css_class)
-        return css_classes
-
-    @classmethod
-    def get_inline_styles(cls, instance):
-        """
-        Returns a dictionary of CSS attributes to be added as style="..." to the current HTML tag.
-        """
-        inline_styles = getattr(cls, 'default_inline_styles', {})
-        css_style = instance.glossary.get('inline_styles')
-        if css_style:
-            inline_styles.update(css_style)
-        return inline_styles
-
-    @classmethod
-    def get_html_tag_attributes(cls, instance):
-        """
-        Returns a dictionary of attributes, which shall be added to the current HTML tag.
-        This method normally is called by the models's property method ``html_tag_ attributes``,
-        which enriches the HTML tag with those attributes converted to a list as
-        ``attr1="val1" attr2="val2" ...``.
-        """
-        attributes = getattr(cls, 'html_tag_attributes', {})
-        return dict((attr, instance.glossary.get(key, '')) for key, attr in attributes.items())
-
-    @classmethod
     def sanitize_model(cls, instance):
         """
         This method is called, before the model is written to the database. It can be overloaded
@@ -359,7 +332,7 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
         """
         Return a representation of the given instance suitable for a serialized representation.
         """
-        return {'glossary': instance.glossary}
+        return {'glossary': instance.glossary, 'pk': instance.pk}
 
     @classmethod
     def add_inline_elements(cls, instance, inlines):
@@ -379,7 +352,7 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
         No child will be removed if wanted_children is smaller than the current number of children.
         """
         from cms.api import add_plugin
-        current_children = parent.get_children().count()
+        current_children = parent.get_num_children()
         for _ in range(current_children, wanted_children):
             child = add_plugin(parent.placeholder, child_class, parent.language, target=parent)
             if isinstance(child_glossary, dict):
@@ -454,8 +427,8 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
         """
         try:
             if obj and obj.parent and obj.position > 0:
-                previnst = obj.parent.get_children().order_by('position')[obj.position - 1]
-                return previnst.get_plugin_instance()
+                prev_inst = obj.parent.get_children().order_by('position')[obj.position - 1]
+                return prev_inst.get_plugin_instance()
         except ObjectDoesNotExist:
             pass
         return None, None
@@ -467,8 +440,8 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
         """
         try:
             if obj and obj.parent:
-                nextinst = obj.parent.get_children().order_by('position')[obj.position + 1]
-                return nextinst.get_plugin_instance()
+                next_inst = obj.parent.get_children().order_by('position')[obj.position + 1]
+                return next_inst.get_plugin_instance()
         except (IndexError, ObjectDoesNotExist):
             pass
         return None, None
@@ -504,8 +477,5 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
         toolbar = getattr(request, 'toolbar', None)
         edit_mode = getattr(toolbar, 'edit_mode', False) and getattr(placeholder, 'is_editable', True)
         if edit_mode:
-            if LooseVersion(cms_version) < LooseVersion('3.4.0'):
-                edit_mode = placeholder.has_change_permission(request)
-            else:
-                edit_mode = placeholder.has_change_permission(request.user)
+            edit_mode = placeholder.has_change_permission(request.user)
         return edit_mode