+++ docs/source/changelog.rst
@@ -4,6 +4,18 @@
 Release History
 ===============
 
+
+0.14
+----
+* Added static rendering of a serialized representation of plugins copied from a ``placeholder``
+  to the clipboard. For details, please read on how to :ref:`strides`.
+
+0.13.1
+------
+* Prepare for Django-1.11 compatibility: Replace renderer classes by specialized widgets
+  overriding its ``render()`` method.
+
+
 0.13
 ----
 * Added Leaflet Plugin which allows to integrate interactive maps from Google, Mapbox and